Rename schema navigation action bar file
