+++ src/vs/workbench/contrib/positronConnections/browser/components/schemaNavigationActionBar.tsx
